Added Python 3 compatibility.
The only thing I had to do was add parentheses around all the 'print'
statements in each source file. I also removed the restriction in setup.py
against installing with Python 3.x. I haven't tested this extensively, but it
seems to work with a LabJack U3 on Ubuntu 15.10 running Python 3.4.